requirements.txt
certifi==2019.3.9
chardet==3.0.4
entrypoints==0.3
flake8==3.7.7
idna==2.8
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
python-dotenv==0.9.1
requests==2.21.0
urllib3==1.24.2
